Refactor: split up main.cpp
- Moved the Event struct and its to_str method to event.cpp and
  event.hpp
- Moved response parsing code into parsing.cpp and parsing.hpp
- Moved request handling to requests.cpp and requests.hpp
